% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_ptype.R
\name{arrow_ptype}
\alias{arrow_ptype}
\title{\code{arrow_ptype} vector}
\usage{
arrow_ptype(...)
}
\arguments{
\item{...}{expressions to turn into arrow data types.}
}
\value{
An S3 vector of class \code{arrow_ptypes}.
}
\description{
This creates a vector of calls that prints anc can be manipulate as a normal S3 vector.
}
