% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_arrow_ptype.R
\name{as_arrow_ptype}
\alias{as_arrow_ptype}
\alias{as_arrow_ptype.Schema}
\alias{as_arrow_ptype.data.frame}
\alias{as_arrow_ptype.FileSystemDataset}
\alias{as_arrow_ptype.Table}
\alias{as_arrow_ptype.StructArray}
\alias{as_arrow_ptype.character}
\alias{as_arrow_ptype.list}
\title{As arrow ptype}
\usage{
as_arrow_ptype(x, .fn = NULL, tbl_name = NULL)

\method{as_arrow_ptype}{Schema}(x, .fn = NULL, tbl_name = NULL)

\method{as_arrow_ptype}{data.frame}(x, .fn = NULL, tbl_name = NULL)

\method{as_arrow_ptype}{FileSystemDataset}(x, .fn = NULL, tbl_name = NULL)

\method{as_arrow_ptype}{Table}(x, .fn = NULL, tbl_name = NULL)

\method{as_arrow_ptype}{StructArray}(x, .fn = NULL, tbl_name = NULL)

\method{as_arrow_ptype}{character}(x, tbl_name = NULL)

\method{as_arrow_ptype}{list}(x, tbl_name = NULL)
}
\description{
As arrow ptype
}
