% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_types_list.R
\name{as_types_list}
\alias{as_types_list}
\alias{as_types_list.default}
\alias{as_types_list.arrow_ptype}
\alias{as_types_list.Schema}
\alias{as_types_list.data.frame}
\alias{as_types_list.FileSystemDataset}
\alias{as_types_list.Table}
\alias{as_types_list.StructArray}
\title{Lists of data types}
\usage{
as_types_list(x, .fn = NULL, ...)

\method{as_types_list}{default}(x, .fn = NULL, ...)

\method{as_types_list}{arrow_ptype}(x, .fn = NULL, ...)

\method{as_types_list}{Schema}(x, .fn = NULL, ...)

\method{as_types_list}{data.frame}(x, .fn = NULL, ...)

\method{as_types_list}{FileSystemDataset}(x, .fn = NULL, ...)

\method{as_types_list}{Table}(x, .fn = NULL, ...)

\method{as_types_list}{StructArray}(x, .fn = NULL, ...)
}
\description{
Lists of data types
}
