% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls_helpers.R
\name{calls_helpers}
\alias{calls_helpers}
\alias{encall}
\alias{call_add_ns}
\title{Call Helpers}
\usage{
encall(x)

call_add_ns(.call, .ns)
}
\arguments{
\item{x}{a call.}

\item{calls}{a list of calls.}
}
\description{
Call Helpers
}
