% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_callables.R
\name{check_callable}
\alias{check_callable}
\alias{check_call}
\alias{check_simple_call}
\alias{check_ns_call}
\alias{check_ns_intetnal_call}
\alias{check_ns_symbol}
\alias{check_symbol}
\alias{check_symbolic}
\alias{check_function}
\alias{check_closure}
\alias{check_formula}
\alias{check_environment}
\title{Checks for calls and callables.}
\usage{
check_callable(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_call(x, ..., arg = rlang::caller_arg(x), call = rlang::caller_env())

check_simple_call(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_ns_call(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_ns_intetnal_call(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_ns_symbol(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_symbol(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_symbolic(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_function(x, ..., arg = rlang::caller_arg(x), call = rlang::caller_env())

check_closure(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_formula(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_environment(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\description{
Checks for calls and callables.
}
