% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_calls_exist.R
\name{check_calls_exist}
\alias{check_calls_exist}
\alias{check_call_exists}
\alias{check_arrow_fn}
\alias{check_arrow_fn.default}
\alias{check_arrow_fn.vcall}
\title{Check that call names exist}
\usage{
check_call_exists(.call, envir = rlang::caller_env(), verbose = FALSE)

check_calls_exist(.calls, envir = rlang::caller_env(), verbose = FALSE)

check_arrow_fn(x)

\method{check_arrow_fn}{default}(x)

\method{check_arrow_fn}{vcall}(x)
}
\arguments{
\item{.call}{a call.}

\item{envir}{an environment.}

\item{verbose}{logical. Defaults to FALSE. If TRUE it prints the function in the form \code{ns::fn()}.}

\item{.calls}{a list of call.}
}
\description{
Check that call names exist
}
