% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_numeric}
\alias{check_numeric}
\title{Check if argument is numeric}
\usage{
check_numeric(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
\code{x} invisibly if is numeric, an error otherwise.
}
\description{
Check if argument is numeric
}
\examples{
check_numeric(letters)
check_numeric(1:10)
}
