% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quosure.R
\name{check_quosure}
\alias{check_quosure}
\title{Check input is a quosure}
\usage{
check_quosure(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{.}

\item{allow_null}{.}

\item{arg}{.}

\item{call}{.}
}
\description{
Check input is a quosure
}
