% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checks}
\alias{checks}
\alias{check_call_simple}
\alias{check_call}
\alias{check_function}
\alias{check_lambda_function}
\alias{check_scalar_character}
\alias{check_scalar_double}
\alias{check_scalar_integer}
\alias{check_scalar_list}
\alias{check_scalar_logical}
\alias{check_scalar_vector}
\alias{check_bool}
\alias{check_string}
\alias{check_character}
\alias{check_double}
\alias{check_integer}
\alias{check_list}
\alias{check_logical}
\alias{check_vector}
\title{Vector checks}
\usage{
check_call_simple(x)

check_call(x)

check_function(x)

check_lambda_function(x)

check_scalar_character(x)

check_scalar_double(x)

check_scalar_integer(x)

check_scalar_list(x)

check_scalar_logical(x)

check_scalar_vector(x)

check_bool(x)

check_string(x)

check_character(x)

check_double(x)

check_integer(x)

check_list(x)

check_logical(x)

check_vector(x)
}
\arguments{
\item{x}{input object.}
}
\value{
an error if the condition is not satisfied. The function returns \code{x} invisibly otherwise.
}
\description{
Vector checks
}
\examples{
input <- 2L
check_scalar_integer(input)
check_character(input)

}
