% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_call.R
\name{eval_call}
\alias{eval_call}
\alias{eval_call.default}
\alias{eval_call.vcall}
\title{TEvaluate calls}
\usage{
eval_call(x, env = parent.env())

\method{eval_call}{default}(x, env = parent.env())

\method{eval_call}{vcall}(x)
}
\arguments{
\item{x}{a list of calls.}

\item{...}{}
}
\description{
TEvaluate calls
}
