% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ns.R
\name{find_ns}
\alias{find_ns}
\alias{.auto_ns}
\alias{auto_ns}
\title{Find Namespaces}
\usage{
.auto_ns(x, full_out = TRUE)

auto_ns(..., ns_only = FALSE)
}
\arguments{
\item{x}{.}

\item{full_out}{.}
}
\description{
Find Namespaces
}
