% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-spaces.R
\name{hex_to_rgb}
\alias{hex_to_rgb}
\title{Convert hexadecimal colours to RGB colour channels.}
\usage{
hex_to_rgb(hex)
}
\arguments{
\item{hex}{A character vector containing hex representations of RGB colours.}
}
\value{
A \code{tibble} of red, green and blue colour channels.
}
\description{
Convert hexadecimal colours to RGB colour channels.
}
\examples{
hex_to_rgb(c("#5f9e3a"))
}
