% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-spaces.R
\name{hsv_to_rgb}
\alias{hsv_to_rgb}
\title{Convert HSV to RGB}
\usage{
hsv_to_rgb(hsv)
}
\arguments{
\item{hsv}{A dataframe or matrix with H, S and V colour channels located in the columns 1 to 3, respectively. H in degrees in [0, 360], S and L in [0, 1]}
}
\value{
A \code{tibble} of red, green and blue colour channels.
}
\description{
Convert HSV to RGB
}
