% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_callable.R
\name{is_callable}
\alias{is_callable}
\alias{is_call}
\alias{is_simple_call}
\alias{is_ns_call}
\alias{is_ns_internal_call}
\alias{is_ns_symbol}
\alias{is_symbol}
\alias{is_symbolic}
\alias{is_function}
\alias{is_closure}
\alias{is_formula}
\alias{is_environment}
\title{Test fr callables}
\usage{
is_callable(x)

is_call(x, name = NULL, n = NULL, ns = NULL)

is_simple_call(x, ns = NULL)

is_ns_call(x, name = "::", n = NULL, ns = NULL)

is_ns_internal_call(x, name = ":::", n = NULL, ns = NULL)

is_ns_symbol(x, ns = NULL, private = NULL)

is_symbol(x, name = NULL)

is_symbolic(x)

is_function(x)

is_closure(x)

is_formula(x, scoped = NULL, lhs = NULL)

is_environment(x)
}
\description{
Test fr callables
}
