% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-spaces.R
\name{lab_to_xyz}
\alias{lab_to_xyz}
\title{Convert from Lab space to XYZ colour channels.}
\usage{
lab_to_xyz(lab)
}
\arguments{
\item{Lab}{A dataframe or matrix with L, a and b colour channels located in the columns 1 to 3, respectively.}
}
\description{
Convert from Lab space to XYZ colour channels.
}
