% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_cnd.R
\name{multi_cnd}
\alias{multi_cnd}
\alias{vec_ptype_abbr.multi_cnd}
\alias{vec_ptype_full.multi_cnd}
\alias{names.multi_cnd}
\alias{names<-.multi_cnd}
\alias{format.multi_cnd}
\alias{obj_print_header.multi_cnd}
\alias{obj_print_data.multi_cnd}
\alias{vec_math.multi_cnd}
\alias{and_c}
\alias{and_c.default}
\alias{and_c.multi_cnd}
\alias{or_c}
\alias{or_c.default}
\alias{or_c.multi_cnd}
\title{\code{multi_cnd} vector}
\usage{
multi_cnd(x, ...)

\method{vec_ptype_abbr}{multi_cnd}(x)

\method{vec_ptype_full}{multi_cnd}(x)

\method{names}{multi_cnd}(x)

\method{names}{multi_cnd}(x) <- value

\method{format}{multi_cnd}(x)

\method{obj_print_header}{multi_cnd}(x)

\method{obj_print_data}{multi_cnd}(x)

\method{vec_math}{multi_cnd}(.fn, x, ...)

and_c(x, ...)

\method{and_c}{default}(...)

\method{and_c}{multi_cnd}(x, ..., verbose = TRUE)

or_c(x, ...)

\method{or_c}{default}(...)

\method{or_c}{multi_cnd}(x, ..., verbose = TRUE)
}
\arguments{
\item{x}{a vector.}

\item{...}{conditions.}
}
\value{
An S3 vector of class \code{multi_cnd}.
}
\description{
This creates a vector of conditions re. a vector.
}
\examples{
multi_cnd(letters, is.numeric, ~ is.character(.x))
}
