% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named-call.R
\name{named-call}
\alias{named-call}
\alias{named_call}
\alias{named_call0}
\alias{as_named_call}
\alias{as_named_call_list}
\title{Create Named (standard) Call}
\usage{
named_call(.fn, ..., .args = NULL, .ns = NULL, .private = FALSE)

named_call0(.fn, .args = NULL)

as_named_call(x)

as_named_call_list(..., .named = FALSE)
}
\description{
Create Named (standard) Call
}
