% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_helpers.R
\name{names_helpers}
\alias{names_helpers}
\alias{re_name}
\alias{re_name2}
\title{Names helpers}
\usage{
re_name(x, .fn = snakecase::to_snake_case, ...)

re_name2(x, .fn = ~.x, ..., .prefix = "x", .pad = "_")
}
\arguments{
\item{x}{a vector or data frame.}

\item{.fn}{a function. Same rules as for \code{purrr::map}.}

\item{...}{arguments passed to \code{.fn}.}

\item{.prefix}{defaults to "x". Prefix to use in case \code{x} has no names.}

\item{.pad}{defaults to "_". Padding to use in case \code{x} has no names .}
}
\description{
Names helpers
}
\examples{
iris_nonames <- unname(iris)
names(re_name(iris))
names(re_name2(iris_nonames))

}
