% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_call.R
\name{namespaced-call}
\alias{namespaced-call}
\alias{ns_call}
\title{Create Namespaced Call}
\usage{
ns_call(.ns = NULL, .fn = NULL, ..., .private = FALSE, .strict = FALSE)
}
\description{
Create Namespaced Call
}
