% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_to_call.R
\name{objs_to_calls}
\alias{objs_to_calls}
\alias{obj_is_call}
\alias{check_obj_is_call}
\alias{objs_are_call}
\alias{check_objs_are_call}
\alias{obj_as_call}
\alias{objs_as_call}
\title{Inputs as calls}
\usage{
obj_is_call(x)

check_obj_is_call(x)

objs_are_call(..., all = TRUE)

check_objs_are_call(...)

obj_as_call(x, ..., .find_ns = TRUE)

objs_as_call(..., .find_ns = TRUE, .named = TRUE)
}
\arguments{
\item{x}{expression or character vector to test as call or transform to call.}
}
\description{
Inputs as calls
}
