% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemr-class.R
\name{palette,schemr-method}
\alias{palette,schemr-method}
\title{Plot the colour palette}
\usage{
\S4method{palette}{schemr}(value)
}
\arguments{
\item{value}{A schemr class object}
}
\value{
No return value, calls a barplot of the colour pallette.
}
\description{
Plot the colour palette
}
