% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemr-class.R
\name{plot,schemr,ANY-method}
\alias{plot,schemr,ANY-method}
\title{Plot the clustered image data}
\usage{
\S4method{plot}{schemr,ANY}(x, y = NULL, ...)
}
\arguments{
\item{x}{A schemr class object}

\item{y}{Not used, NULL}

\item{...}{Other arguments to pass onto `plot`}
}
\value{
No return value, calls a raster plot of the clustered image data.
}
\description{
Plot the clustered image data
}
