% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quos_helpers.R
\name{quos_helpers}
\alias{quos_helpers}
\alias{quos_set_expr}
\alias{quos_get_env}
\alias{quos_set_env}
\title{Get/set quosures expressions and environments}
\usage{
quos_set_expr(x, exprs)

quos_get_env(x)

quos_set_env(x, env)
}
\arguments{
\item{x}{quosures.}

\item{exprs}{list of expressions.}

\item{env}{an environment.}
}
\description{
Get/set quosures expressions and environments
}
