% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-spaces.R
\name{rgb_to_hsv}
\alias{rgb_to_hsv}
\title{Convert RGB to HSV}
\usage{
rgb_to_hsv(rgb)
}
\arguments{
\item{rgb}{A dataframe or matrix with red, green and blue colour channels located in the columns 1 to 3, respectively. Colour channel values should be between 0 and 255, inclusive.}
}
\value{
A \code{tibble} of H, S and V colour channels.
}
\description{
Convert RGB to HSV
}
