% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemr-class.R
\docType{class}
\name{schemr-class}
\alias{schemr-class}
\title{Create the schemr class, which holds the palette and image data}
\description{
Create the schemr class, which holds the palette and image data
}
\section{Fields}{

\describe{
\item{\code{image}}{An array of dimension (Image width) by (Image height) by (3 colour channels) that contains the data of the original image}

\item{\code{clustered_image}}{An array of dimension (Image width) by (Image height) by (3 colour channels) that contains the data of the image with clustered colour blocks}

\item{\code{palette}}{A character vector that contains the colours of the resulting colour palette}
}}

\section{Methods}{

\describe{
\item{\code{print(x)}}{Print the colour palette.}
}}

