% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemr-package.R
\docType{package}
\name{schemr-package}
\alias{schemr-package}
\title{schemr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Filippo DallOlio \email{fdallolio@brocku.ca}

}
\keyword{internal}
