% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_call.R
\name{split_call}
\alias{split_call}
\alias{split_calls}
\title{Split calls into call name, args, and namespace}
\usage{
split_call(x)

split_calls(...)
}
\arguments{
\item{x}{a call.}

\item{...}{calls.}
}
\description{
Split calls into call name, args, and namespace
}
