% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helpers.R
\name{str_numpad}
\alias{str_numpad}
\title{Pad number strings}
\usage{
str_numpad(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
a character vector of numbers padded left with '0'.
}
\description{
Pad number strings
}
\examples{
str_numpad(1:15)
}
