% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helpers.R
\name{v_names}
\alias{v_names}
\title{Vector of integer characters padded with \code{V}}
\usage{
v_names(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
a character vector of numbers padded left with 'V' or 'V0'.
}
\description{
Vector of integer characters padded with \code{V}
}
\examples{
v_names(letters)
}
