% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcall.R
\name{vcall}
\alias{vcall}
\alias{is_vcall}
\alias{format.vcall}
\alias{vec_ptype_full.vcall}
\alias{vec_ptype_abbr.vcall}
\alias{obj_print_data.vcall}
\alias{obj_print_footer.vcall}
\alias{names.vcall}
\alias{names<-.vcall}
\alias{[<-.vcall}
\alias{[[.vcall}
\alias{[[<-.vcall}
\title{\code{vcall} vector}
\usage{
vcall(..., ns = NULL)

is_vcall(x)

\method{format}{vcall}(x)

\method{vec_ptype_full}{vcall}(x)

\method{vec_ptype_abbr}{vcall}(x)

\method{obj_print_data}{vcall}(x)

\method{obj_print_footer}{vcall}(x)

\method{names}{vcall}(x)

\method{names}{vcall}(x) <- value

\method{[}{vcall}(x, ..., values) <- value

\method{[[}{vcall}(x, i)

\method{[[}{vcall}(x, i) <- value
}
\arguments{
\item{...}{expressions to turn into calls.}
}
\value{
An S3 vector of class \code{vcall}.
}
\description{
This creates a vector of calls that prints anc can be manipulate as a normal S3 vector.
}
\examples{
vcall(mean, "mean", mean(), "mean()")
}
