% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_slice.R
\name{vec_slice}
\alias{vec_slice}
\alias{vec_head}
\alias{vec_tail}
\title{Vector Slicing}
\usage{
vec_slice(x, i, ..., error_call = rlang::current_env())

vec_head(x, n)

vec_tail(x, n)
}
\arguments{
\item{x}{.}

\item{i}{.}

\item{...}{.}

\item{error_call}{.}

\item{n}{.}
}
\description{
Vector Slicing
}
