% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls_helpers.R
\name{calls_helpers}
\alias{calls_helpers}
\alias{call_name2}
\alias{calls_nms}
\alias{call_args2}
\alias{calls_args}
\alias{call_ns2}
\alias{calls_ns}
\alias{call_match2}
\alias{calls_match}
\alias{call_set_ns}
\alias{calls_set_ns}
\alias{call_text}
\alias{call_text.default}
\alias{call_self_match}
\alias{call_set_name}
\alias{call_no_ns}
\alias{call_has_ns}
\alias{call_expand}
\alias{call_base}
\alias{encall}
\title{Get names/args/ns (and set ns) of calls lists.}
\usage{
call_name2(calls)

calls_nms(calls)

call_args2(calls)

calls_args(calls)

call_ns2(calls)

calls_ns(calls)

call_match2(calls, env = rlang::caller_env(), ...)

calls_match(calls, env = parent.env(), ...)

call_set_ns(x, value)

calls_set_ns(calls, ns)

call_text(x, unname = TRUE, shrink = TRUE)

\method{call_text}{default}(x, unname = TRUE, shrink = TRUE)

call_self_match(
  call,
  ...,
  defaults = !is.null(call_ns(call)),
  dots_env = NULL,
  dots_expand = TRUE
)

call_set_name(x, value)

call_set_ns(x, value)

call_no_ns(x)

call_has_ns(x)

call_expand(...)

call_base(name, ns = NULL)

encall(x)
}
\arguments{
\item{calls}{a list of calls.}

\item{x}{a call.}
}
\description{
Get names/args/ns (and set ns) of calls lists.
}
