% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls_helpers.R
\name{calls_helpers}
\alias{calls_helpers}
\alias{call_name2}
\alias{calls_nms}
\alias{call_args2}
\alias{calls_args}
\alias{call_ns2}
\alias{calls_ns}
\alias{call_match2}
\alias{calls_match}
\alias{call_set_ns}
\alias{calls_set_ns}
\alias{call_text}
\alias{call_text.default}
\alias{call_text.vcall}
\title{Get names/args/ns (and set ns) of calls lists.}
\usage{
call_name2(calls)

calls_nms(calls)

call_args2(calls)

calls_args(calls)

call_ns2(calls)

calls_ns(calls)

call_match2(calls, env = parent.env(), ...)

calls_match(calls, env = parent.env(), ...)

call_set_ns(x, ns)

calls_set_ns(calls, ns)

call_text(x, unname = TRUE, shrink = TRUE)

\method{call_text}{default}(x, unname = TRUE, shrink = TRUE)

\method{call_text}{vcall}(x, unname = TRUE, shrink = TRUE)
}
\arguments{
\item{calls}{a list of calls.}

\item{x}{a call.}
}
\description{
Get names/args/ns (and set ns) of calls lists.
}
