% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemr-package.R
\docType{package}
\name{schemr-package}
\alias{schemr-package}
\title{schemr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-dallolio/schemr}
  \item Report bugs at \url{https://github.com/f-dallolio/schemr/issues}
}

}
\author{
\strong{Maintainer}: Filippo DallOlio \email{fdallolio@brocku.ca}

}
\keyword{internal}
