% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-spaces.R
\name{xyz_to_lab}
\alias{xyz_to_lab}
\title{Convert from XYZ colour channels to Lab space.}
\usage{
xyz_to_lab(xyz)
}
\arguments{
\item{xyz}{A dataframe or matrix with X, Y and Z colour channels located in the columns 1 to 3, respectively.}
}
\value{
A \code{tibble} of L, a and b colour space values.
}
\description{
Convert from XYZ colour channels to Lab space.
}
\examples{
x <- sample(x = 40:60, size = 10, replace = TRUE)
y <- sample(x = 40:60, size = 10, replace = TRUE)
z <- sample(x = 40:60, size = 10, replace = TRUE)
xyz_to_lab(data.frame(x = x, y = y, z = z))
}
